pyproject.toml
[tool.poetry]
name = "bert_as_a_service"
version = "0.1.0"
description = ""
authors = ["Bin Wen <dracifer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.4"
tensorflow = "1.14"
bert-serving-server = "^1.10.0"
bert-serving-client = "^1.10.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
whitenoise = {extras = ["brotli"], version = "^5.2.0"}
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
